+++ nata-monkey/src/main/java/com/nata/monkeys/QLearningMonkey.java
@@ -3,9 +3,8 @@ package com.nata.monkeys;
 import com.nata.action.Action;
 import com.nata.action.ActionFactory;
 import com.nata.action.BackAction;
-import com.nata.action.SwipeAction;
 import com.nata.cmd.AdbDevice;
-import com.nata.element.UINode;
+import com.nata.element.Widget;
 import com.nata.state.State;
 
 import java.util.*;
@@ -21,17 +20,19 @@ public class QLearningMonkey extends AbstractMonkey {
 
     // <state,action,value>
     private Map<State, Map<Action, Double>> QMap = new HashMap<>();
-    private Set<String> activitySet =  new HashSet<>();
+    private Set<String> activitySet = new HashSet<>();
     private List<Action> actionList = new ArrayList<>();
     private Action lastAction = null;
     private Action backAction = null;
 
     //final variables
     private final int ACTION_COUNTS = 50;
-    private final double GAMA= 0.7;
+    private final double ALPHA = 0.2;
+    private final double GAMA = 0.2;
     private final double PUNISH_OUT_PACKAGE = -10.0;
-    private final double PUNISH_SWIPE_NOTWORK= - 0.5;
+//    private final double PUNISH_SWIPE_NOTWORK= - 0.5;
 
+    List<Action> chosenActions= new ArrayList<>();
 
     public QLearningMonkey(String pkg, String act, AdbDevice device) {
         super("QLearningMonkey", pkg, act, device);
@@ -40,30 +41,46 @@ public class QLearningMonkey extends AbstractMonkey {
         backAction = actionFactory.createBackAction();
     }
 
-    private Action chooseActionFromTable(Map<Action, Double> actionTable) {
-        Action choosedAction = null;
+    private Action chooseActionFromState(State curState) {
+        Map<Action, Double> actionTable = QMap.get(curState);
+        if(actionTable == null || actionTable.size() == 0){
+            System.out.println("Error cannot get actionTable or no action can be choosed");
+            return backAction;
+        }
+
+        System.out.println("[ActionList] " + actionTable);
+
+        Action chosenAction = null;
         double charming_max = 0.0;
+
         for (Map.Entry<Action, Double> entry : actionTable.entrySet()) {
             double value = entry.getValue();
             if (value > charming_max) {
-                choosedAction = entry.getKey();
+                chosenAction = entry.getKey();
                 charming_max = value;
+                chosenActions.clear();
+                chosenActions.add(chosenAction);
+            }else if(Math.abs(value - charming_max) < 0.0001){ //almost equal
+                chosenAction = entry.getKey();
+                chosenActions.add(chosenAction);
             }
         }
-        return choosedAction;
+        Collections.shuffle(chosenActions);
+        System.out.println("charming_max:" + charming_max);
+        return chosenActions.get(0);
     }
 
-    private double getRewardFromExperience(State curState,Action action, State nextState){
+    private double getRewardFromExperience(State curState, Action action, State nextState) {
         double reward = 0.0;
         // punish to run out of package
-        if(!nextState.getAppPackage().equals(getPkg())){
-             reward += PUNISH_OUT_PACKAGE;
+        if (!nextState.getAppPackage().equals(getPkg())) {
+            reward += PUNISH_OUT_PACKAGE;
         }
 
-        // if the swipe action makes the state not change then give punish
-        if(action instanceof SwipeAction && curState.equals(nextState)){
-             reward += PUNISH_SWIPE_NOTWORK;
-        }
+//        // if the swipe action makes the widget not change then give punish
+//        if(action instanceof SwipeAction && curState.equals(nextState)){
+//             reward += PUNISH_SWIPE_NOTWORK;
+//        }
 
         // the action reward
         reward += action.getReward();
@@ -71,63 +88,101 @@ public class QLearningMonkey extends AbstractMonkey {
     }
 
 
-    private double getMaxValueOfState(State nextState){
+    private double getMaxValueOfState(State nextState) {
         Map<Action, Double> actionTable = QMap.get(nextState);
-        double reward = 0.0;
-        for (Double value: actionTable.values()) {
-           if(value > reward) {
-               reward = value;
-           }
+        double maxValue = 0.0;
+        for (Double value : actionTable.values()) {
+            if (value > maxValue) {
+                maxValue = value;
+            }
         }
-        return reward;
+        return maxValue;
     }
 
 
-    private void updateValue(State curState,Action action, State nextState){
+    private void updateValue(State curState, Action action, State nextState) {
         Map<Action, Double> actionTable = QMap.get(curState);
-        double reward = getRewardFromExperience(curState,action,nextState);
+        double reward = getRewardFromExperience(curState, action, nextState);
         double future_reward = getMaxValueOfState(nextState);
         actionTable.put(action, reward + GAMA * future_reward);
     }
 
-    @Override
-    public void play() {
+
+    private void startApp() {
         System.out.println("QLearning Monkey start playing...");
         clearAppData();
         System.out.println("cleaning app data...");
+
         restartAction.fire();
         actionList.add(restartAction);
         System.out.println("starting app success...");
+    }
+
+    /**
+     * Add the state to the value map of Q and add the new activity to activitySet
+     * @param state
+     */
+    private void addStateToQMap(State state){
+        //add to activitySet
+        if (isInCurrentPkg()) {
+            activitySet.add(state.getActivity());
+        }
+        //add to QMap
+        if (QMap.get(state) == null) {
+            Map<Action, Double> table = actionFactory.getActionsFromState(state);
+            QMap.put(state, table);
+            getDevice().screenShot(state.hashCode() + "");
+        }
+    }
+
+    /**
+     * help the monkey to get back to app
+     * @return the current state
+     */
+    private State getBackToApp() {
+        while(!isInCurrentPkg()){
+            // if monkey get out of pkg because of back action
+            if (lastAction instanceof BackAction) {
+                restartAction.fire();
+                actionList.add(restartAction);
+            } else {
+                backAction.fire();
+                actionList.add(backAction);
+                if(!isInCurrentPkg()){
+                    backAction.fire();
+                    actionList.add(backAction);
+                }
+                lastAction = backAction;
+            }
+        }
+        // get current state
+        State state = getCurrentState();
+        addStateToQMap(state);
+        return state;
+    }
+
+
+
+
+    @Override
+    public void play() {
+        startApp();
 
         // initial state
-        State curState = getCurrentState();
-        System.out.println(curState);
-        activitySet.add(curState.getActivity());
+        State curState= getCurrentState();
+        addStateToQMap(curState);
 
-        if (QMap.get(curState) == null) {
-            Map<Action, Double> table= actionFactory.getActionsFromState(curState);
-            QMap.put(curState, table);
-        }
         // repeat select -> execute -> update
         int cnt = 0;
         while ((++cnt) < ACTION_COUNTS) {
-            Action action;
-            //select
+            // if not in pkg, get it back
             if(!isInCurrentPkg()) {
-                // if monkey get out of pkg because of back action
-                if (lastAction instanceof BackAction) {
-                    action = restartAction;
-                } else {
-                    action = backAction;
-                }
-            }else{
-                Map<Action, Double> actionTable = QMap.get(curState);
-                System.out.println("[ActionList] " +actionTable);
-                action = chooseActionFromTable(actionTable);
+                curState = getBackToApp();
+                continue;
             }
 
-            //check if monkey can use his knowledge and change the action
-            brainStorm(action);
+            //select action
+            Action action = chooseActionFromState(curState);
 
             //execute
             System.out.println(action);
@@ -138,19 +193,10 @@ public class QLearningMonkey extends AbstractMonkey {
             //update
             State nextState = getCurrentState();
             System.out.println(nextState);
+            addStateToQMap(nextState);
+            updateValue(curState, action, nextState);
 
-            //add to activitySet
-            if(isInCurrentPkg()){
-                activitySet.add(nextState.getActivity());
-            }
-
-            if (QMap.get(nextState) == null) {
-                Map<Action, Double> table = actionFactory.getActionsFromState(nextState);
-                QMap.put(nextState, table);
-            }
-            updateValue(curState,action,nextState);
             curState = nextState;
-
             System.out.println("---------------------------[Experience]---------------------------");
         }
     }
@@ -164,29 +210,29 @@ public class QLearningMonkey extends AbstractMonkey {
     public void report() {
         //report the activities reached
         System.out.println("--------------------[Activities report]--------------------");
-        for (String activity: activitySet) {
-           System.out.println(activity);
+        for (String activity : activitySet) {
+            System.out.println(activity);
         }
 
         System.out.println("--------------------[Action report]--------------------");
         int actionCount = actionList.size();
-        System.out.println("Actions count: "+actionCount);
-        for (Action action: actionList) {
+        System.out.println("Actions count: " + actionCount);
+        for (Action action : actionList) {
             System.out.println(action);
         }
 
         System.out.println("--------------------[Widget report]--------------------");
-        Set<UINode> widgetSet = getWidgetSet();
-        int widgetCount= widgetSet.size();
-        System.out.println("Widget count: "+widgetCount);
-        for (UINode node: widgetSet) {
+        Set<Widget> widgetSet = getWidgetSet();
+        int widgetCount = widgetSet.size();
+        System.out.println("Widget count: " + widgetCount);
+        for (Widget node : widgetSet) {
             System.out.println(node);
         }
 
         System.out.println("--------------------[State report]--------------------");
         int stateCount = QMap.size();
-        System.out.println("State count: "+stateCount);
-        for (State state: QMap.keySet()) {
+        System.out.println("State count: " + stateCount);
+        for (State state : QMap.keySet()) {
             System.out.println(state);
         }
     }